+++ src/dipdup/indexes/tezos_token_transfers/fetcher.py
@@ -1,6 +1,5 @@
 from __future__ import annotations
 
-import logging
 from typing import TYPE_CHECKING
 
 from dipdup.indexes.tezos_tzkt import TezosTzktFetcher
@@ -16,6 +15,7 @@ class TokenTransferFetcher(TezosTzktFetcher[TezosTokenTransferData]):
 
     def __init__(
         self,
+        name: str,
         datasources: tuple[TezosTzktDatasource, ...],
         token_addresses: set[str],
         token_ids: set[int],
@@ -24,8 +24,12 @@ class TokenTransferFetcher(TezosTzktFetcher[TezosTokenTransferData]):
         first_level: int,
         last_level: int,
     ) -> None:
-        super().__init__(datasources, first_level, last_level)
-        self._logger = logging.getLogger('dipdup.fetcher')
+        super().__init__(
+            name=name,
+            datasources=datasources,
+            first_level=first_level,
+            last_level=last_level,
+        )
         self._token_addresses = token_addresses
         self._token_ids = token_ids
         self._from_addresses = from_addresses