+++ src/dipdup/indexes/tezos_tzkt.py
@@ -27,11 +27,13 @@ class TezosTzktFetcher(
 ):
     def __init__(
         self,
+        name: str,
         datasources: tuple[TezosTzktDatasource, ...],
         first_level: int,
         last_level: int,
     ) -> None:
         super().__init__(
+            name=name,
             datasources=datasources,
             first_level=first_level,
             last_level=last_level,